moved usermodel to a own map and cleaned up the code
