+++ packages/backend/src/db.ts
@@ -1,15 +1,4 @@
-import {UserItem} from "@webshop-app/shared"
-import { connect, model, Schema } from "mongoose"
-
-const UserSchema = new Schema({
-    fullName: String,
-    email: String,
-    password: String,
-    phoneNumber: Number,
-    address: String
-})
-
-const UserModel = model<UserItem>("User", UserSchema)
+import { connect } from "mongoose"
 
 export const setupMongoDb = async (url: string) => {
     try {
@@ -20,9 +9,3 @@ export const setupMongoDb = async (url: string) => {
         throw e;
     }
 }
-
-export const saveUser = async (user: UserItem): Promise<UserItem | null> => {  
-    const newModel = new UserModel(user);
-    newModel.save();
-    return newModel;
-  };
